refactor(yacfg): Refactor yacfg_batch_cli.py, cli_arguments.py, yacfg_batch.py,  for improved code organization and readability
- Update and remove unused import statements
- Created separate functions for boolize, split_key_value, and parse_key_value_list
- Added type annotations to function parameters and return types
- Reorganized code to improve clarity and maintainability
