+++ src/yacfg/cli_arguments.py
@@ -1,23 +1,62 @@
-# Copyright 2018 Red Hat Inc.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#    http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
 import argparse
 import json
 import re
+from typing import Dict, List, Tuple, Union
 
-from . import __version__
-from .meta import NAME, DESCRIPTION
+from . import DESCRIPTION, NAME, __version__
+
+REX_BOOL_TRUE = re.compile(r"^(([Tt][Rr][Uu][Ee])|[Yy][Ee][Ss]|(1)|[Oo][Nn])$")
+REX_BOOL_FALSE = re.compile(r"^(([Ff][Aa][Ll][Ss][Ee])|([Nn][Oo])|(0)|[Oo][Ff][Ff])$")
+
+
+def boolize(data: str) -> Union[bool, None]:
+    """Convert input values from string to bool if possible
+
+    :param data: input data from argparse
+    :type data: str
+    :return: converted boolean value, or None if not specified
+    :rtype: bool or None
+
+    :raises ValueError: if conversion failed
+    """
+    if data is None:
+        return None
+    if REX_BOOL_TRUE.match(data):
+        return True
+    if REX_BOOL_FALSE.match(data):
+        return False
+    raise ValueError(f"Cannot convert input option {data} to bool")
+
+
+def split_key_value(item: str) -> Tuple[str, str]:
+    """Split KEY=VALUE specifier or raise a ValueError exception.
+
+    :param item: input KEY=VALUE string option.
+    :type item: str
+
+    :raises ValueError: if cannot split
+
+    :return: pair of split key and value
+    :rtype: str, str
+    """
+    try:
+        key, value = item.split("=", 1)
+    except ValueError:
+        raise ValueError(f'Missing KEY=VALUE pair in option "{item}"')
+    return key, value
+
+
+def parse_key_value_list(items: List[Dict[str, str]]) -> Dict[str, str]:
+    """Convert a list of dictionaries into a single dictionary.
+
+    :param items: List of dictionaries.
+    :type items: list[dict]
+
+    :return: Merged dictionary with key-value pairs from all dictionaries.
+    :rtype: dict
+    """
+    return {k: v for item in items for k, v in item.items() if isinstance(item, dict)}
+
 
 parser = argparse.ArgumentParser(
     prog="{} {}".format(NAME, __version__),
@@ -168,57 +207,4 @@ group_misc.add_argument(
     "--version", help="Display version information", action="store_true"
 )
 
-REX_BOOL_TRUE = re.compile(r"^(([Tt][Rr][Uu][Ee])|[Yy][Ee][Ss]|(1))$")
-REX_BOOL_FALSE = re.compile(r"^(([Ff][Aa][Ll][Ss][Ee])|([Nn][Oo])|(0))$")
-
-
-def boolize(data):
-    """Convert input values from string to bool if possible
-
-    :param data: input data from argparse
-    :type data: str
-    :return: converted boolean value, or None if not specified
-    :rtype: bool or None
-
-    :raises ValueError: if conversion failed
-    """
-    if data is None:
-        return None
-    if REX_BOOL_TRUE.match(data):
-        return True
-    if REX_BOOL_FALSE.match(data):
-        return False
-    raise ValueError(f'Cannot convert input option {data} to bool')
-
-
-def split_key_value(item):
-    """Split KEY=VALUE specifier or raise an ValueError exception.
-
-    :param item: input KEY=VALUE string option.
-    :type item: str
-
-    :raises ValueError: if cannot split
-
-    :return: pair of split key and value
-    :rtype: str, str
-    """
-    try:
-        key, value = item.split("=", 1)
-    except ValueError:
-        raise ValueError('Missing KEY=VALUE pair in option "{}"'.format(item))
-    return key, value
-
-
-def parse_key_value_list(items):
-    """Split all KEY=VALUE items in a list of such options.
-
-    :param items: list of KEY=VALUE string options to be split.
-    :type items: list[str]
-
-    :raises ValueError: if any option cannot be split.
-
-    :return: a map of KEY: VALUE split options
-    :rtype: dict
-    """
-    result = dict([split_key_value(item) for item in items])
-    return result
+args = parser.parse_args()
