+++ src/yacfg_batch/yacfg_batch.py
@@ -1,17 +1,3 @@
-# Copyright 2018 Red Hat Inc.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#    http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
 from __future__ import print_function
 
 import copy
@@ -21,6 +7,7 @@ import os
 import yaml
 
 import yacfg.yacfg
+
 from .exceptions import YacfgBatchException
 
 LOG = logging.getLogger(__name__)
