+++ src/yacfg_batch/yacfg_batch_cli.py
@@ -1,24 +1,10 @@
-# Copyright 2018 Red Hat Inc.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#    http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
 from __future__ import print_function
 
 import logging
 import sys
 
-from yacfg import logger_settings
-from yacfg.meta import NAME
+from yacfg import NAME, logger_settings
+
 from . import __version__
 from .cli_arguments import parser
 from .yacfg_batch import generate
